Update list of dependencies with thorough instructions
Please note that `awk`, `tr`, `sort`, `head`, `cp`, `find`, and `cut` are separate command-line utilities that are typically available in most Linux/Unix distributions and are not built-in commands in Bash.

Since it is safe to assume such commands will be available by default there is still the likely chance that they are not depending on the environment. As such we check for them anyways.